corrected decimal types

+++ back-end/database/migrations/2023_07_31_014207_create_nutrition_facts_per100_g_s_table.php
@@ -13,13 +13,13 @@ return new class extends Migration
     {
         Schema::create('Health.NutritionFactsPer100G', function (Blueprint $table) {
             $table->id('NutritionFactsID');
-            $table->integer('Calories');
-            $table->integer('Fats');
-            $table->integer('Collesterol');
-            $table->integer('Sodium');
-            $table->integer('Fiber');
-            $table->integer('Carbs');
-            $table->integer('Protein');
+            $table->decimal('Calories', 7, 2);
+            $table->decimal('Fats', 7, 2);
+            $table->decimal('Collesterol', 7, 2);
+            $table->decimal('Sodium', 7, 2);
+            $table->decimal('Fiber', 7, 2);
+            $table->decimal('Carbs', 7, 2);
+            $table->decimal('Protein', 7, 2);
         });
     }
